requirements.txt
# 📦 Discord Bot 主要依賴套件
discord.py>=2.3.2        # 🧩 Discord API 操作核心套件

# 📒 資料管理套件
PyYAML>=6.0.1            # 🧩 YAML 讀寫套件，統一資料儲存格式

# 🌍 時區與時間處理
pytz>=2024.1             # 🧩 時區管理，保證時間正確顯示

# 🖥️ 開發輔助工具
watchdog>=4.0.0          # 🧩 檔案監控模組，支援模組熱重載

# 🔮 預留擴充套件（可視需求啟用）
aiohttp>=3.9.1           # 🧩 HTTP 請求，未來如 API 抓取使用
python-dotenv>=1.0.1     # 🧩 環境變數管理，保持設定檔乾淨
